Default to using libc on Unix and WASI.
Switch is-terminal on Unix and WASI to use `libc` instead of `rustix` by
default. This matches the `IsTerminal` implementation in std, and avoids
pulling in all of rustix as a dependency just for one function.

This retains rustix as an option, which may be enabled by using
`default-features = false` and enabling either the `rustix-std` or
`rustix-no_std` features.