Port rust-lang/rust#119664 to is-terminal.
Port rust-lang/rust#119664 to is-terminal. The commit message is:

Fix msys2 tty detection for /dev/ptmx

Our "true negative" detection assumes that if at least one std handle is a Windows console then no other handle will be a msys2 tty pipe. This turns out to be a faulty assumption in the case of  `/dev/ptmx`.